Fix Bug: No Error When Deleting an Empty Like
If the user had no likes, the server could not
respond to a POST to the delete page. Checking
that the like exists after querying for it before
trying to delete it was necessary.

+++ blog.py
@@ -189,7 +189,8 @@ class UnlikeHandler(Handler):
 		user = self.valid_user_cookie()
 		post = Post.get_by_id(int(post_id), parent=blog_key())
 		like = Like.gql("WHERE author = :author AND post = :post", author = user, post = post).get()
-		like.delete()
+		if like:
+			like.delete()
 		self.redirect('/post/%s' % post_id)
 
 app = webapp2.WSGIApplication([('/', IndexHandler),
